+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Broken plots caused by single quotes in NCBI taxon names.
+
 ## [v2.5.0]
 ### Added
 - `--kraken_confidence` to specify a threshold score.